+++ mood_math.h
@@ -7,7 +7,7 @@
 #define P3 (3*M_PI)/2
 
 float fixAng(float a);
-float degToRad(int a);
+float degToRad(float a);
 float dist(float ax, float ay, float bx, float by, float ang);
 
 #endif
